feat: Display years in the countdown if required
Since the number of years to the final data were been obtained, the view
has been modified to display that information if required. The number of
years left will only be displayed when there's at least one remaining.
If the year count is zero, it won't show up in the view.