fix: Initialize countdown at component mount
When the application started, the countdown values were all set to zero
during a few moments. The cause for this error was that the counter
values were not been initialized until the 'setInterval' called its
parameter function for the first time, which meant the countdown values
will be unitialized for a second.

The current change ensures that the countdown is property initialized
before the user gets to see the page, so they don't experience a delay
from when the page loads to the time in which they're presented with the
real countdown.